fix : (검색 브런치 이긴한데)OCR 버그 수정함... + 카메라 권한 위치 사진등록 눌렀을때로 변경
