+++ app/src/main/java/com/blackcows/butakaeyak/ui/home/HomeFragment.kt
@@ -1,6 +1,9 @@
 package com.blackcows.butakaeyak.ui.home
 
+import android.Manifest
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -10,7 +13,10 @@ import android.view.ViewGroup
 import android.view.animation.Animation
 import android.view.animation.AnimationUtils
 import android.widget.Toast
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.ViewModel
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.blackcows.butakaeyak.MainViewModel
 import com.blackcows.butakaeyak.R
@@ -27,11 +33,14 @@ import com.blackcows.butakaeyak.ui.schedule.TimeToGroup
 import com.blackcows.butakaeyak.ui.search.SearchFragment
 import com.blackcows.butakaeyak.ui.take.fragment.TakeAddFragment
 import com.blackcows.butakaeyak.ui.textrecognition.OcrFragment
+import com.blackcows.butakaeyak.ui.textrecognition.OcrFragment.Companion
 import com.blackcows.butakaeyak.ui.viewmodels.MyGroupViewModel
 import com.blackcows.butakaeyak.ui.viewmodels.UserViewModel
+import dagger.hilt.android.scopes.ViewModelScoped
 import io.ktor.util.date.WeekDay
 import kotlinx.coroutines.delay
 import java.time.LocalDate
+import kotlin.coroutines.suspendCoroutine
 
 
 private const val TAG = "HomeFragment"
@@ -58,6 +67,17 @@ class HomeFragment : Fragment(), View.OnClickListener {
 
     private var isSpread: Boolean = false
 
+    //카메라 권한
+    private val REQUEST_CODE_PERMISSIONS = 10
+    private val REQUIRED_PERMISSIONS =
+        mutableListOf (
+            Manifest.permission.CAMERA,
+            Manifest.permission.RECORD_AUDIO
+        ).apply {
+            if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.P) {
+                add(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+            }
+        }.toTypedArray()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -179,6 +199,10 @@ class HomeFragment : Fragment(), View.OnClickListener {
         }
     }
 
+    private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all {//카메라 권한 관련
+        ContextCompat.checkSelfPermission(requireActivity(), it) == PackageManager.PERMISSION_GRANTED
+    }
+
     private fun fragmentInit () {
         // 나랑 가까운 약국
         binding.homeConnection1.setOnClickListener{
@@ -194,7 +218,12 @@ class HomeFragment : Fragment(), View.OnClickListener {
         }
         // 사진 등록
         binding.btnAddMedicine2.setOnClickListener{
-            MainNavigation.addFragment(OcrFragment(), FragmentTag.OCRFragment)
+
+            if (allPermissionsGranted()) {
+                MainNavigation.addFragment(OcrFragment(), FragmentTag.OCRFragment)
+            }else{
+                ActivityCompat.requestPermissions(requireActivity(), REQUIRED_PERMISSIONS, REQUEST_CODE_PERMISSIONS)
+            }
         }
     }
